fix: move FleetSemaphore to fleet/ layer to preserve cascade narrowing
FleetSemaphore belongs in fleet/ (L2), not server/ (L3). Fleet test
files must import from autoskillit.fleet, not autoskillit.server._factory,
to avoid breaking the server→fleet file-level cascade narrowing enforced
by test_test_filter_cascade.py.

- Add fleet/_semaphore.py with FleetSemaphore class
- Re-export FleetSemaphore from fleet/__init__.py
- Remove FleetSemaphore from server/_factory.py; import from fleet lazily
- Update all test imports to use autoskillit.fleet
- Revert _test_filter.py server entry to original (no new file-level entries)
