+++ src/autoskillit/fleet/__init__.py
@@ -8,6 +8,7 @@ from ._api import execute_dispatch
 from ._liveness import is_dispatch_session_alive
 from ._prompts import _build_food_truck_prompt as _build_food_truck_prompt
 from ._prompts import _build_l2_sous_chef_block as _build_l2_sous_chef_block
+from ._semaphore import FleetSemaphore
 from .result_parser import L2ParseResult, parse_l2_result_block
 from .sidecar import (
     IssueSidecarEntry,
@@ -78,6 +79,7 @@ __all__ = [
     "write_captured_values",
     "write_initial_state",
     "is_dispatch_session_alive",
+    "FleetSemaphore",
     "IssueSidecarEntry",
     "append_sidecar_entry",
     "compute_remaining_issues",
